Add files via upload
Thrilled to share that I secured 2nd Place 🥈 in the Career Code Hackathon conducted by the Meta Developer Communities (MDC) at GITAM University on 14th December 2024!

It was an incredible experience to compete alongside brilliant minds and tackle challenging problems that pushed the boundaries of creativity and innovation.

Congratulations to all the winners —
🏆 1st Place: Sarika Sai Vardhan
🥉 3rd Place: Arjit Kumar

A big thank you to Team MDC GITAM for organizing such an inspiring and well-executed event. Grateful for the opportunity and excited to keep building, learning, and growing with this amazing community! 🚀💻

#Hackathon #MDC #GITAM #MetaDeveloperCommunities #CareerCodeHackathon #Innovation #ProblemSolving #Gratitude #LinkedInWins